Fix division of long tensors for Python3.5

+++ torchero/meters/binary_scores.py
@@ -162,19 +162,19 @@ class TPMeter(BaseMeter):
 
     @staticmethod
     def _recall(tp, tn, fp, fn):
-        return torch.where(tp+fn==0, torch.zeros_like(tp, dtype=torch.float), (tp/(tp + fn)).float())
+        return torch.where(tp+fn==0, torch.zeros_like(tp, dtype=torch.float), tp.float()/(tp + fn).float())
 
     @staticmethod
     def _precision(tp, tn, fp, fn):
-        return torch.where(tp+fp==0, torch.zeros_like(tp, dtype=torch.float), (tp/(tp + fp)).float())
+        return torch.where(tp+fp==0, torch.zeros_like(tp, dtype=torch.float), tp.float()/(tp + fp).float())
 
     @staticmethod
     def _specificity(tp, tn, fp, fn):
-        return torch.where(tn+fn==0, torch.zeros_like(tp, dtype=torch.float), (tn/(tn + fp)).float())
+        return torch.where(tn+fn==0, torch.zeros_like(tp, dtype=torch.float), tn.float()/(tn + fp).float())
 
     @staticmethod
     def _npv(tp, tn, fp, fn):
-        return torch.where(tn+fn==0, torch.zeros_like(tp, dtype=torch.float), (tn/(tn + fn)).float())
+        return torch.where(tn+fn==0, torch.zeros_like(tp, dtype=torch.float), tn.float()/(tn + fn).float())
 
     def _gen_fbeta(self, beta=1):
         def fbeta(tp, tn, fp, fn):
